requirements.txt
requests==2.31.0
openai==2.7.1
python-dotenv==1.0.0
pydantic==2.6.1
